move all auth related stuff to auth package
