+++ ebics-rest-api/src/main/kotlin/org/ebics/client/ebicsrestapi/auth/InMemoryAuthProviderConfiguration.kt
@@ -1,4 +1,4 @@
-package org.ebics.client.ebicsrestapi
+package org.ebics.client.ebicsrestapi.auth
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Configuration
