+++ ebics-rest-api/src/main/kotlin/org/ebics/client/ebicsrestapi/auth/key/ApiKeyConfiguration.kt
@@ -1,4 +1,4 @@
-package org.ebics.client.ebicsrestapi.key
+package org.ebics.client.ebicsrestapi.auth.key
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
