+++ ebics-rest-api/src/main/kotlin/org/ebics/client/ebicsrestapi/auth/key/ApiKeyProperties.kt
@@ -1,4 +1,4 @@
-package org.ebics.client.ebicsrestapi.key
+package org.ebics.client.ebicsrestapi.auth.key
 
 
 import org.springframework.boot.context.properties.ConfigurationProperties
@@ -10,9 +10,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties
  */
 
 @ConfigurationProperties(prefix = "ebics.api")
-data class ApiKeyProperties(
-    val clients: Map<String, ApiKey> = emptyMap()
-) {
+data class ApiKeyProperties(val clients: Map<String, ApiKey> = emptyMap()) {
     data class ApiKey(
         val key: String,
         val role: String = "api"
