+++ ebics-rest-api/src/main/kotlin/org/ebics/client/ebicsrestapi/auth/ldap/LdapSearchProperties.kt
@@ -1,4 +1,4 @@
-package org.ebics.client.ebicsrestapi.ldap
+package org.ebics.client.ebicsrestapi.auth.ldap
 
 
 import org.springframework.boot.context.properties.ConfigurationProperties
